init empty dir (#116)
* Re-arrange init dir check logic

Why bother with bare when it comes to folder creation? bare seems to be about what to do inside the dir.
Then in the create template, use listdir to validate if dir empty or not.

* Switch to distutils copy_tree

shutil version barfs if the destination already exists. defaults of copy_tree match previous settings.

* version bump